% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_holes_igraph.R
\name{structural_holes_igraph}
\alias{structural_holes_igraph}
\title{structural_holes_igraph}
\usage{
structural_holes_igraph(gs)
}
\arguments{
\item{gs}{igraph object}
}
\value{
Dataframe of strucutral hole measures
}
\description{
This function calulates structural holes measures for all nodes in the network.Input network is a igraph object.
}
