% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_holes_matrix_function.R
\name{structural_holes_matrix}
\alias{structural_holes_matrix}
\title{structural_holes_matrix}
\usage{
structural_holes_matrix(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
Dataframe of strucutral hole measures
}
\description{
This function calulates structural holes measures for all nodes in the network.Input network is a matrix.
}
