% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_holes_network_function.R
\name{structural_holes_network}
\alias{structural_holes_network}
\title{structural_holes_network}
\usage{
structural_holes_network(net)
}
\arguments{
\item{net}{network object}
}
\value{
Dataframe of strucutral hole measures
}
\description{
This function calulates structural holes measures for all nodes in the network.Input network is a network object.
}
